changed comment at end of program
Changed the comment precededing the "if INITIAL_PRODUCTION" that encloses the workbook.save line that writes the assessment_schedule.xls file. I changed the comment only to test that the repository is functioning, because my initial attempt to push to this repository failed because of a network error. So, I want to make sure that the repository isn't corrupt by making this trivial commit.